.history/requirements_20240601122731.txt
selenium==4.4.3
mysql-connector-python==8.0.30
psycopg2-binary
requests
beautifulsoup4
html5lib
